fixup! Move completion to adapter
